Also check if the user is the same in the servlet.

+++ portfolio/src/main/java/com/google/sps/servlets/DeleteServlet.java
@@ -18,6 +18,9 @@ import com.google.appengine.api.datastore.Query;
 import com.google.appengine.api.datastore.Query.SortDirection;
 import com.google.appengine.api.datastore.Key;
 import com.google.appengine.api.datastore.KeyFactory;
+import com.google.appengine.api.users.UserService;
+import com.google.appengine.api.users.UserServiceFactory;
+import com.google.appengine.api.datastore.EntityNotFoundException;
 
 /** Servlet that deletes some data. */
 @WebServlet("/delete-data")
@@ -26,9 +29,19 @@ public class DeleteServlet extends HttpServlet {
   @Override
   public void doPost(HttpServletRequest request, HttpServletResponse response) throws IOException {
       DatastoreService datastore = DatastoreServiceFactory.getDatastoreService();
-      long id = Long.parseLong(request.getParameter("comment-id"));
-      Key key = KeyFactory.createKey("Comment", id);
-      datastore.delete(key);
+      long commentId = Long.parseLong(request.getParameter("comment-id"));
+
+      Key key = KeyFactory.createKey("Comment", commentId);
+      String currentUserId = UserServiceFactory.getUserService().getCurrentUser().getUserId();
+      String commentUserId = null;
+      try {
+          commentUserId = (String) datastore.get(key).getProperty("userId");
+      } catch (EntityNotFoundException e) {
+          commentUserId = null;
+      }
+
+      if (commentUserId.equals(currentUserId))
+        datastore.delete(key);
 
       response.sendRedirect("/index.html");
   }
